refactor: move all parquet-associated features into parquet-writer mod
